requirements.txt
Flask
google-cloud-vision
openpyxl
gunicorn